Move `test_pagination_blackbox` to a separate class and add logging.
